requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==3.2.2
langchain==0.1.16
flask
pypdf
python-dotenv
-e .